Create Tracker
This file is written in AHK. If you don't have a way to implement AHK as an executable, or run this in a debugging and editing tool like SciTE4AHK, it won't be much use to you. None of this is proprietary. Some of it is (I think) clever, but it's all stuff anyone could have written to meet the needs I needed to meet at the time. This is the main portion of a full time and production tracking program. The other portions involve .bat scripts, additional executables (including a launcher for this program), and version controls and user checks that will prevent just anyone from using it without heavy modification first. It's only here to demonstrate my ability to write in AHK.